Updated TypeAheadWithOther component and moved it under Form. Updated styleguide for TypeAheadWithOther
