+++ styles/_forms.scss
@@ -14,8 +14,8 @@
 }
 
 .form-row {
+  margin: 20px 0;
   min-height: 85px;
-  padding-bottom: 20px;
 }
 
 .two-item-row {
